requirements_l.txt
TA-Lib==0.4.19
cython
requests
gunicorn
gevent
pandas
python-socketio
websocket-client
apscheduler
redis


